cmd: test_enter_suspend: Add test_enter_suspend command
Add test_enter_suspend command using PSCI.
This command is only to test suspend sequence from U-Boot.
The resume will not work.

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>

Series-to: u-boot@lists.denx.de
Series-cc: gregory.clement@bootlin.com
Series-cc: thomas.petazzoni@bootlin.com
Series-cc: thomas.richard@bootlin.com
Series-cc: nm@ti.com
Series-cc: u-kumar1@ti.com
Cover-letter:
Suspend to RAM support for K3 J7200
This series is the U-Boot part of the work to add the suspend to RAM
support for the K3 J7200 EVM board.

During the boot R5 SPL makes a copy of DM-Firmware in memory.
Resume detection is done by reading a magic value in a pmic register
(set by DM-Firmware).

If a resume is detected, R5 SPL run the exit retention sequence of the
DDR. Then it restores TF-A in SRAM (it saved itself to DDR), and load
DM-Firmware using the copy done during the boot (fit image processing
is skipped).

The copy of TF-A/DM-Firmware, the SPL stack and malloc are located in a
reserved memory region (for the kernel point of view) to avoid any
memory corruption.

A test_enter_sleep command was implemented to be able to test the
suspend sequence from U-Boot. It is only for test, the resume will not
work.
END