arm: mach-k3: j7200: Skip fit processing when resuming
No need to process fit image during resume.
All needed parts are already available in DRAM.

Based on the work of Gregory CLEMENT <gregory.clement@bootlin.com>

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>

Series-version: 3
Series-to: u-boot@lists.denx.de
Series-cc: gregory.clement@bootlin.com
Series-cc: thomas.petazzoni@bootlin.com
Series-cc: thomas.richard@bootlin.com
Series-cc: nm@ti.com
Series-cc: u-kumar1@ti.com
Cover-letter:
Suspend to RAM support for K3 J7200
This series is the U-Boot part of the work to add the suspend to RAM
support for the K3 J7200 EVM board.

During the boot R5 SPL makes a copy of DM-Firmware and TF-A in memory.
Resume detection is done by reading a magic value in a pmic register
(set by DM-Firmware).

If a resume is detected, R5 SPL run the exit retention sequence of the
DDR. Then it load TF-A and DM-Firmware using the copies done during the boot
(fit image processing is skipped).
Before to start TF-A, R5 SPL writes a magic value in scratchpad ram. This
will be used by TF-A to detect that the board is resuming.

The copy of TF-A/DM-Firmware, the SPL stack and malloc are located in a
reserved memory region (for the kernel point of view) to avoid any
memory corruption.

This version is mostly to test the firewall protection with the suspend
to ram.
Some comments (for the v2) were not fixed in this version.
This series has been tested with the series [1] to enable the firewall.
At the end of the resume sequence, TF-A is well protected by the
firewall, but OP-TEE remains unprotected.

[1] https://lore.kernel.org/all/20231229-binman-firewalling-v7-0-47ed4af303fe@ti.com/

END

Cover-changes: 3
- Link to v2:
  https://lore.kernel.org/u-boot/20231107161802.855154-1-thomas.richard@bootlin.com/

Cover-changes: 2
- Remove the test_enter_suspend command
- Link to v1:
  https://lore.kernel.org/u-boot/20231016141135.325698-1-thomas.richard@bootlin.com/