board: ti: j721e: During resume spl notify tf-a that the board is resuming
During the boot a copy of DM-Firmware and TF-A is done in a reserved memory
area.
When resuming, R5 SPL uses these copies instead of the fit image.
R5 SPL writes a magic value in the scratchpad ram to notify TF-A that
the board is resuming.

Based on the work of Gregory CLEMENT <gregory.clement@bootlin.com>

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>

Series-changes: 3
- At resume, R5 SPL doesn't restore TF-A anymore.
  TF-A is started like during a cold boot. R5 SPL will notify that the board is
  resuming using a magic value written in the scratchpad ram.
  TF-A will restore itself.

Series-changes: 2
- Check if TF-A is running in DRAM, if yes no need to restore it
- Remove BL31_START macro, and get TF-A start address from the fit image