j7200: yaml: disable L3 cache
For IO retention support we need to use TIFS 09.00.03, because a change
in TIFS 09.00.06 causes an issue on the IO retention sequence of
DM-Firmware. And there is no way to wakeup the SoC.

But in TIFS 09.00.03, there is no L3 cache support. So it's not possible
to flush it from DM-Firmware during suspend stage.

Disable L3 cache to avoid any issue.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>