ram: k3-ddrss: add exit retention support for j784s4
No big differences with j7200.
I noticed only two differences, which should not be there as the enter
retention sequence is the same.
We will try without them. I added them in #if 0 as a reminder in case of
issue.

Only DDR accesses were checked, I didn't checked yet the PMIC accesses.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>