configs: add defconfigs for the am625-lp-sk
The am62x-lp-sk is a package and reference board spin of the am62x-sk to
showcase the low-power features of the am62x SoC family. Because it so
closely resembles the am62x-sk board, use the preprocessor to inherit
its configuration making the needed changes for this board where
necessary.

Reviewed-by: Dhruva Gole <d-gole@ti.com>
Link: https://lore.kernel.org/u-boot/20240429-am62q-wip-v3-2-17dcfdb1075a@ti.com/
Signed-off-by: Bryan Brattlof <bb@ti.com>