arm: dts: k3-am62-lp: remove CONFIG_OF_UPSTREAM
The v2024.04 version of U-Boot does not have support for the
CONFIG_OF_UPSTREAM option. This requires us to sync device tree source
files to the traditional way to keep the dtbs in sync for this backport.

So copy over the am62x configuration from ti-linux-6.6.y when it was at
13c8fd2a35d5f ("rm64: dts: ti: k3-j784s4-main: add clock and pd to gpu")

Signed-off-by: Bryan Brattlof <bb@ti.com>