arm: dts: k3-*: Add memory node to SPL stages
Add the bootph-all property to the memory node so that it can be
accessed by FDT functions at SPL stage.

The SPL stage requires the memory node to be able to initialize and set
the size of the DRAM banks. For this purpose, make sure all the memory
nodes are present and standardized, and modify them if not.

Signed-off-by: Neha Malcom Francis <n-francis@ti.com>