arm: dts: k3-j722s-evm-u-boot: Use eFuse MAC Address for CPSW3G Port 1
The am65-cpsw-nuss driver supports reading the eFuse in order to fetch
the MAC Address corresponding to CPSW3G Port 1. This requires using the
"ti,syscon-efuse" property in the device-tree with the phandle to the
eFuse registers in the CTRL_MMR region.

The Linux device-tree patch at:
Link: https://lore.kernel.org/r/20240604104425.3770037-1-s-vadapalli@ti.com/
which enables fetching MAC Address from eFuse registers for AM62P is
also applicable to J722S SoC which reuses the CPSW3G Ethernet Switch
present on AM62P SoC.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>