arm: mach-k3: common: subtract CONFIG_SYS_MEM_TOP_HIDE
To keep the number of DDR capacity boards spin down we can use the
CONFIG_SYS_MEM_TOP_HIDE option to hide an arbitrary amount or RAM from
U-Boot and the kernel it will boot. However when enabling the caches
this size reduction is not accounted for.

Subtract CONFIG_SYS_MEM_TOP_HIDE from the effective memory size before
we enable the instruction and data caches.

Signed-off-by: Bryan Brattlof <bb@ti.com>