arm: mach-k3: am625_fdt: fixup memory{} node for different capacities
Depending on the board and it's variants, there are a significant number
of instances where there could be multiple DDR capacities that need to
be supported. Because U-Boot has initialized DDR in previous boot stages
it knows the most about the capacity of it.

Rather than creating a device-tree board file for every board variant
with DDR capacity changes, have U-Boot fixup the kernel device tree with
the capacity of the DDR available for the kernel.

Signed-off-by: Bryan Brattlof <bb@ti.com>