HACK: board: ti: am62px: add legacy boot scripting
The AM62Px family of SoCs was initially written to be as distribution
agnostic as possible. This clear contract between the bootloader and the
distribution requires some form of configuration, via EFI variables or a
extlinux.conf file, to describe to the bootloader how to boot the
distribution.

Unfortunately our test distribution and our infrastructure is not setup
for this and requires this information be compiled into U-Boot to
operate properly. For this reason I'm labeling this custom boot script
addition to U-Boot's environment for the board as a hack.

Signed-off-by: Bryan Brattlof <bb@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>