HACK: arm: dts: k3-j7200-common-u-boot: Add overrides for Hyperflash functionality
Linux MMIO reg-mux driver supports the mux functionality when
the parent device is not a syscon, but the U-Boot MMIO mux driver
does not currently support it and thus functionality is broken
when device tree is synced with the parent node of a mmio-mux
not being a syscon. To restore functionality override the compatible
in U-boot, the hack can be dropped when ideal solution of porting
the MMIO mux driver changes from Linux to U-Boot is complete.
Also while at it add the bootphase marking for nodes necessary for
Hyperflash boot.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Acked-by: Andrew Davis <afd@ti.com>