HACK: arm: dts: k3-j721e-common-proc-board-u-boot: Fix SERDES0 configuration
SERDES0 has to be configured for the PCIe + QSGMII multilink configuration
at U-Boot. Linux uses PCIe with SERDES0 Lane 0 while either EthFw or Linux
can use QSGMII with SERDES0 Lane 1. Hence, update the SERDES0 device-tree
nodes to support the same.

Mark this as a HACK since the Sierra SERDES driver expects PLL_CMNLC and
PLL_CMNLC1 as the child clocks of SERDES0, due to which the clocks are
moved to the child nodes corresponding to the individual SERDES links.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>