HACK: Revert "board: ti: j721e: Drop SERDES PHY init from board file"
This reverts commit 434f84b7498335510cb11fae4d0b64c3570ddc9a.

While the correct way to initialize the SERDES PHY is to register the
SERDES drivers with the UCLASS framework, since SERDES was functional
at U-Boot prior to the commit, the commit introduces a regression.
Hence, revert the commit until the SERDES driver is updated to be
registered with the UCLASS framework.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>