HACK: arm: dts: k3-am69-sk-u-boot|k3-j784s4-evm-u-boot: Add overrides for SERDES functionality
Linux MMIO reg-mux driver supports the mux functionality when
the parent device is not a syscon, but the U-Boot MMIO mux driver
does not currently support it and thus functionality is broken
when device tree is synced with the parent node of a mmio-mux
not being a syscon. To restore functionality override the compatible
in U-boot, the hack can be dropped when ideal solution of porting
the MMIO mux driver changes from Linux to U-Boot is complete.

Fixes: 42c73efe15d2 ("arm: dts: k3-j784s4-main: Sync with Linux 6.10-rc5 for SERDES and WIZ")
Cc: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Udit Kumar <u-kumar1@ti.com>
Acked-by: Siddharth Vadapalli <s-vadapalli@ti.com>