HACK: arm: dts: k3-j721e-*-u-boot.dtsi: Add overrides for SERDES
Linux MMIO reg-mux driver supports the mux functionality when the parent
device is not a syscon, but the U-Boot MMIO mux driver does not and thus
SERDES functionality will break once the device tree is synced as the
parent device is not a syscon. Override the compatible in U-Boot, this
hack patch can be dropped once the MMIO mux driver changes have been
ported from Linux to U-Boot.

Cc: Udit Kumar <u-kumar1@ti.com>
Cc: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Neha Malcom Francis <n-francis@ti.com>