HACK: arm: dts: k3-j784s4-*: k3-am69-*: Refactor cdns,phy-mode property
The cdns,phy-mode property is needed for OSPI PHY calibration that
achieves higher clock speeds (greater than 25 MHz DDR) with OSPI.
However, this flag is not used in kernel to enable calibration (it
relies on ospi.phypattern partition being present) and its binding is
not present. Hence, hack around this for the DT sync by moving the
property to the -u-boot.dtsi and -r5-sk.dts for j784s4 and am69
respectively.

Cc: Vaishnav Achath <vaishnav.a@ti.com>
Cc: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Acked-by: Neha Malcom Francis <n-francis@ti.com>