HACK: arm: dts: k3-j7200-common-proc-board-u-boot: Add overrides for SERDES
Linux MMIO reg-mux driver supports the mux functionality when
the parent device is not a syscon, but the U-Boot MMIO mux driver
does not currently support it and thus functionality is broken
when device tree is synced with the parent node of a mmio-mux
not being a syscon. To restore functionality override the compatible
in U-boot, the hack can be dropped when ideal solution of porting
the MMIO mux driver changes from Linux to U-Boot is complete.

Fixes: e98c89973caf ("arm: dts: j7200: DT sync with ti-linux-6.6.y")
Cc: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Aniket Limaye <a-limaye@ti.com>
Acked-by: Siddharth Vadapalli <s-vadapalli@ti.com>