arm: dts: k3-am62-lp-sk-u-boot: Enable usb0_phy_ctrl for USB DFU boot
Add the "bootph-all" property to the "usb0_phy_ctrl" node which is used
by the USB0 Subsystem in order to obtain reference to the USB0_PHY_CTRL
register and determine the frequency of the input reference clock. This is
required for USB DFU boot. While at it, add comments to indicate all USB
DFU dependencies.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>