j784s4_spl.h: add infos for saving/reloading DM
During the boot a copy of DM-Firmware and TF-A is done in a reserved memory
area.
When resuming, R5 SPL uses these copies instead of the fit image.
R5 SPL writes a magic value in the scratchpad ram to notify TF-A that
the board is resuming.

Based on the work of Gregory CLEMENT <gregory.clement@bootlin.com>

Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>