lpddr4: split exit from retention sequence in 3 parts
On J7200 and J784s4, the DDR resume sequence is:
- exit DDR from retention
- de-assert the DDR_RET pin
- restore DDR max frequency
- exit DDR from low power

On J7200, it could be done in one function since it has only one DDR.
But for j784s4, as it has 4 DDRs with 4 controllers, it can't be done
that way.
We have to separate each action in order to de-assert the DDR_RET pins at
the right time, because the DDR_RET pins are all tied together in
hardware.

Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>